+++ cmd/empire/server.go
@@ -59,7 +59,7 @@ func runServer(c *cli.Context) {
 	}
 
 	if c.String(FlagCustomResourcesQueue) != "" {
-		p := newCloudFormationCustomResourceProvisioner(db, ctx)
+		p := newCloudFormationCustomResourceProvisioner(e, ctx)
 		log.Printf("Starting CloudFormation custom resource provisioner")
 		go p.Start()
 	}
@@ -81,8 +81,8 @@ func newServer(c *Context, e *empire.Empire) http.Handler {
 	return middleware.Handler(c, h)
 }
 
-func newCloudFormationCustomResourceProvisioner(db *empire.DB, c *Context) *cloudformation.CustomResourceProvisioner {
-	p := cloudformation.NewCustomResourceProvisioner(db.DB.DB(), c)
+func newCloudFormationCustomResourceProvisioner(e *empire.Empire, c *Context) *cloudformation.CustomResourceProvisioner {
+	p := cloudformation.NewCustomResourceProvisioner(e, c)
 	p.QueueURL = c.String(FlagCustomResourcesQueue)
 	p.Context = c
 	return p